Move date component to utils and reuse it in update queue
